Adding option to exclude courses from enrollment
Added a new parameter in the enroll_student_in_program method to feed in whether to exclude any courese from the enrollment (e.g Careers). Added condition to check whether the course code is in the list that is passed into the function and continue to next iteration if it is. Set the new parameter to an empty list by default