Refactored lms_breadcrumbs
Pivoted breadcrumbs script to have a columns for each aspect of the block. Added separate columns for module, section, lesson, unit, component and populated with None if the type does not warrant it (e.g. module only has the module field populated and section, lesson, unit and component are empty). Added a extra information and normalized the block type to our definitions (e.g. unit instead of vertical)