Updated breadcrumbs logic
Changed harvest_course_tree_new to work with a list of dict as output instead of a massive dict which enables easier work with pandas.
Created a logic whereby module, section, lesson, unit and component are only populated if applicable (e.g. if the block is a module, the rest of the fields will be None)
Created function like dict.get() which safely extracts an index from a list and returns None if the index is out of bounds.